+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Please add to the relevant subsections under Unreleased below on every PR where 
 - `run.log_artifact()` no longer blocks other data uploads until the artifact upload finishes (@timoffex in https://github.com/wandb/wandb/pull/8466)
   - NOTE: Not fixed if using `wandb.require("legacy-service")`
 - Fix Unknown image format error when uploading a gif through tensorboard. (@jacobromero in https://github.com/wandb/wandb/pull/8476)
+- Fix `OSError` from calling `Artifact.add_file` with file paths on mounted filesystems (@tonyyli-wandb in https://github.com/wandb/wandb/pull/8473)
 
 ### Changed